+++ CHANGELOG.md
@@ -3,17 +3,45 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
-## [1.x] - Unreleased
+
+[//] : # ## [1.x] - Unreleased
+
+## [1.6.2] - 2019-07-24
+
+### New features:
+
+    - Add possibility to generate cylinder box from pdb
+    - Add coordinate wrapping to XTC reporter
+
+### Bug fixes:
+
+    - Fix bug restarting MD simulation with constraints
+    - Fix bug that crashed when constraining HIS
+    - Fix colorbar range for plotAdaptive script
+    - Fix bug in HIS protonation
+
+### Behaviour changes from previous version:
+
+    - Allow PELE equilibration to run without a box
+    - Read PELE and MD reports seemingly
+    - Add postprocessing option to simulationRunner
+    - Add proper topology support to selectOnPlot and backtrackAdaptiveTrajectory 
+    - Write topologies.pkl at the start of the first epoch
+
+## [1.6.1] - 2019-04-26
 
 ### New features:
 
     - Add cylindrical box for MD
     - Add script to generate cylindrical box for visualization
+    - Introduce new postprocessing function for MD
 
 ### Bug fixes:
 
     - Fix bug selecting topolgies for odd number of initial conditions
     - Fix bug extracting indexes from topology in MSMClustering
+    - Fix bug in atomset writing PDBs with gaps, now if there are residues with
+      a difference in more than one in the residue number a TER is written 
 
 ### Behaviour changes from previous version: